requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
requests